refactor: remove OneShotCommandBar and preserve Assistant infrastructure
- Remove OneShotCommandBar component and one-shot command execution
- Strip AppAgentService to config and API key management only
- Refactor appAgentStore to keep only config/API key state for Assistant
- Remove IPC channels for one-shot execution (runOneShot, cancel, dispatch)
- Preserve action dispatch IPC for Assistant tool calling
- Move AGENT_ACCESSIBLE_ACTIONS to actionTools.ts for Assistant use
- Update menu entry from "AI Command Bar" to "Canopy Assistant"
- Update settings copy to reference Assistant panel instead of command bar
- Remove agent.commandBar action and keybinding definitions
- Clean up type exports to remove one-shot specific types
