+++ electron/ipc/channels.ts
@@ -247,14 +247,11 @@ export const CHANNELS = {
   COMMANDS_EXECUTE: "commands:execute",
   COMMANDS_GET_BUILDER: "commands:get-builder",
 
-  APP_AGENT_RUN_ONE_SHOT: "app-agent:run-one-shot",
   APP_AGENT_GET_CONFIG: "app-agent:get-config",
   APP_AGENT_SET_CONFIG: "app-agent:set-config",
   APP_AGENT_HAS_API_KEY: "app-agent:has-api-key",
   APP_AGENT_TEST_API_KEY: "app-agent:test-api-key",
   APP_AGENT_TEST_MODEL: "app-agent:test-model",
-  APP_AGENT_CANCEL: "app-agent:cancel",
-  APP_AGENT_DISPATCH_ACTION: "app-agent:dispatch-action",
 
   // Assistant channels
   ASSISTANT_SEND_MESSAGE: "assistant:send-message",
