+++ shared/types/actions.ts
@@ -207,8 +207,7 @@ export type ActionId =
   | "notes.create"
   | "notes.delete"
   | "notes.reveal"
-  | "devServer.start"
-  | "agent.commandBar";
+  | "devServer.start";
 
 export interface ActionContext {
   projectId?: string;
