+++ shared/types/keymap.ts
@@ -100,7 +100,6 @@ export type KeyAction =
   | "agent.terminal"
   | "agent.focusNextWaiting"
   | "agent.focusNextFailed"
-  | "agent.commandBar"
 
   // Assistant panel
   | "assistant.open"
