--- src/components/AppAgent/index.ts
@@ -1,1 +1,0 @@
-export { OneShotCommandBar, default as OneShotCommandBarDefault } from "./OneShotCommandBar";
