+++ src/store/index.ts
@@ -74,4 +74,3 @@ export { useTwoPaneSplitStore } from "./twoPaneSplitStore";
 export type { TwoPaneSplitConfig } from "./twoPaneSplitStore";
 
 export { useAppAgentStore, cleanupAppAgentStore } from "./appAgentStore";
-export type { CommandBarStatus } from "./appAgentStore";
